+++ src/structure/dataframe.rs
@@ -5,7 +5,7 @@
 //! * There are various ways to declare `DataFrame`
 //!     * `DataFrame::new()` - Empty dataframe
 //!     * `DataFrame::with_header(header: Vec<&str>)` - Empty dataframe with header
-//!     * `DataFrame::from_matrix(header: Vec<&str>, mat: Matrix)` - from matrix
+//!     * `DataFrame::from_matrix(mat: Matrix)` - from matrix
 //! 
 //! ### Example
 //!
@@ -71,7 +71,7 @@
 //!     a["z"] = c!(6);
 //!
 //!     // Get
-//!     assert_eq!(&c!(1,2,3), a.get("x"));
+//!     assert_eq!(&c!(1,2,3), a.get("x").unwrap());
 //!     assert_eq!(c!(4,5), a["y"]);
 //! }
 //! ```
@@ -83,13 +83,13 @@
 //! | Command | Mean [s] | Min [s] | Max [s] | Relative |
 //! |:---|---:|---:|---:|---:|
 //! | `./target/release/df_bench_csv --features dataframe` | 1.027 ± 0.016 | 0.998 | 1.056 | 27.0 |
-//! | `./target/release/df_bench_cdf --features dataframe` | 0.038 ± 0.002 | 0.035 | 0.042 | 1.0 |
+//! | `./target/release/df_bench_nc --features dataframe` | 0.038 ± 0.002 | 0.035 | 0.042 | 1.0 |
 //!
 //! * Benchmark codes are as follows.
 //!
 //! ### Example (Benchmark codes)
 //!
-//! #### 1. CDF 
+//! #### 1. netcdf 
 //!
 //! ```rust
 //! extern crate peroxide;
@@ -105,7 +105,9 @@
 //!     df.write_nc("example_data/df_bench.nc")?;
 //!
 //!     // read_nc(path: &str, header: Vec<&str>)
-//!     let dg = DataFrame::read_nc("example_data/df_bench.nc", vec!["x", "y", "z"])?;
+//!     let dg = DataFrame::read_nc("example_data/df_bench.nc")?;
+//!     // or read_nc with specific header
+//!     // let dg = DataFrame::read_nc_by_header("example_data/df_bench.nc", vec!["x", "z"])?;
 //!     dg.print();
 //!
 //!     Ok(())
@@ -156,6 +158,13 @@ pub struct DataFrame {
     pub data: IndexMap<String, Vec<f64>>,
 }
 
+impl PartialEq for DataFrame {
+    fn eq(&self, other: &Self) -> bool {
+        self.data == other.data
+    }
+}
+
+/// Pretty view for DataFrame
 impl fmt::Display for DataFrame {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "{}", self.spread())
@@ -166,41 +175,50 @@ impl Index<&str> for DataFrame {
     type Output = Vec<f64>;
 
     fn index(&self, index: &str) -> &Self::Output {
-        self.get(index)
+        self.get(index).unwrap()
     }
 }
 
 impl IndexMut<&str> for DataFrame {
     fn index_mut(&mut self, index: &str) -> &mut Self::Output {
-        self.data.get_mut(index).unwrap()
+        match self.data.get_mut(index) {
+            Some(mut v) => v,
+            None => panic!("There are no corresponding value")
+        }
     }
 }
 
 #[allow(unused_parens)]
 impl DataFrame {
+    /// Declare empty DataFrame
     pub fn new() -> Self {
         DataFrame {
             data: IndexMap::new()
         }
     }
 
+    /// Declare empty DataFrame with header
     pub fn with_header(header: Vec<&str>) -> Self {
         let l = header.len();
-        Self::from_matrix(
-            header,
-            Matrix {
-                data: vec![],
-                row: 0,
-                col: l,
-                shape: Col,
-            }
-        )
+        let mut data = IndexMap::with_capacity(l);
+        for i in 0 .. l {
+            data.insert(header[i].to_string(), vec![]);
+        }
+        DataFrame {
+            data
+        }
     }
 
+    pub fn len(&self) -> usize {
+        self.data.len()
+    }
+
+    /// Insert key & value pair (or only value)
     pub fn insert(&mut self, key: &str, value: Vec<f64>) {
         self.data.insert(key.to_owned(), value);
     }
 
+    /// Insert row
     pub fn insert_row(&mut self, value: Vec<f64>) {
         assert_eq!(self.data.len(), value.len());
         for (v, val) in self.data.values_mut().zip(value) {
@@ -208,14 +226,28 @@ impl DataFrame {
         }
     }
 
-    pub fn get(&self, head: &str) -> &Vec<f64> {
-        &self.data.get(head).unwrap()
+    /// Get value by ref
+    pub fn get(&self, head: &str) -> Option<&Vec<f64>> {
+        self.data.get(&head.to_string())
     }
 
+    /// Get iterator of headers
     pub fn headers(&self) -> Keys<String, Vec<f64>> {
         self.data.keys()
     }
 
+    pub fn set_header(&mut self, header: Vec<&str>) {
+        for i in 0 .. header.len() {
+            match self.data.get_index_mut(i) {
+                Some((mut k, _)) => {
+                    *k = header[i].to_string();
+                }
+                None => panic!("New header is longer than original header")
+            }
+        }
+    }
+
+    /// Convert to matrix
     pub fn to_matrix(&self) -> Matrix {
         let mut data: Vec<f64> = vec![];
         let mut r = 0usize;
@@ -232,14 +264,16 @@ impl DataFrame {
         matrix(data, r, c, Col)
     }
 
-    pub fn from_matrix(header: Vec<&str>, mat: Matrix) -> Self {
+    /// Convert from matrix
+    pub fn from_matrix(mat: Matrix) -> Self {
         let mut df: DataFrame = DataFrame::new();
         for i in 0 .. mat.col {
-            df.insert(header[i], mat.col(i));
+            df.insert(format!("{}", i).as_str(), mat.col(i));
         }
         df
     }
 
+    /// For pretty print
     pub fn spread(&self) -> String {
         let r: usize = self.data.values().fold(0, |val, v2| max(val, v2.len()));
 
@@ -377,11 +411,13 @@ impl DataFrame {
     
 }
 
+/// To deal with CSV file format
 pub trait WithCSV: Sized {
     fn write_csv(&self, file_path: &str) -> Result<(), Box<dyn Error>>;
     fn read_csv(file_path: &str, delimiter: char) -> Result<Self, Box<dyn Error>>;
 }
 
+/// CSV with DataFrame (inefficient)
 impl WithCSV for DataFrame {
     fn write_csv(&self, file_path: &str) -> Result<(), Box<dyn Error>> {
         let mut wtr = WriterBuilder::new().from_path(file_path)?;
@@ -411,12 +447,7 @@ impl WithCSV for DataFrame {
         let headers_vec = rdr.headers()?;
         let headers = headers_vec.iter().map(|x| x).collect::<Vec<&str>>();
         let l = headers.len();
-        let mut result = DataFrame::from_matrix(headers, Matrix {
-            data: vec![], 
-            row: 0, 
-            col: l, 
-            shape: Col
-        });
+        let mut result = DataFrame::with_header(headers);
 
         for rec in rdr.deserialize() {
             let record: HashMap<String, String> = rec?;
@@ -432,11 +463,14 @@ impl WithCSV for DataFrame {
     }
 }
 
+/// To deal with NetCDF file format
 pub trait WithNetCDF: Sized { 
     fn write_nc(&self, file_path: &str) -> Result<(), Box<dyn Error>>;         
-    fn read_nc(file_path: &str, header: Vec<&str>) -> Result<Self, Box<dyn Error>>;
+    fn read_nc(file_path: &str) -> Result<Self, Box<dyn Error>>;
+    fn read_nc_by_header(file_path: &str, header: Vec<&str>) -> Result<Self, Box<dyn Error>>;
 }                                                                               
                                                                                 
+/// NetCDF with DataFrame (efficient)
 impl WithNetCDF for DataFrame {         
     fn write_nc(&self, file_path: &str) -> Result<(), Box<dyn Error>> {
         let mut f = netcdf::create(file_path)?;
@@ -454,16 +488,27 @@ impl WithNetCDF for DataFrame {
 
         Ok(())
     }
-    fn read_nc(file_path: &str, header: Vec<&str>) -> Result<Self, Box<dyn Error>> {
+
+    fn read_nc(file_path: &str) -> Result<Self, Box<dyn Error>> {
+        let f = netcdf::open(file_path)?;
+        let mut df = DataFrame::new();
+        for (k, v) in f.root.variables.iter() {
+            let data: Vec<f64> = v.get_double(false)?;
+            df.insert(k, data);
+        }
+        Ok(df)
+    }
+
+    fn read_nc_by_header(file_path: &str, header: Vec<&str>) -> Result<Self, Box<dyn Error>> {
         let f = netcdf::open(file_path)?;
         let mut df = DataFrame::with_header(header.clone());
         for k in header {
-            let var = match f.root.variables.get(k) {
+            let val = match f.root.variables.get(k) {
                 Some(v) => v,
-                None => panic!("There are no corresponding data"),
+                None => panic!("There are no corresponding values"),
             };
-            let data: Vec<f64> = var.get_double(false)?;
-            df.insert(k, data);
+            let data: Vec<f64> = val.get_double(false)?;
+            df[k] = data;
         }
         Ok(df)
     }